+++ simple-jwt-login/src/Modules/Settings/AuthCodesSettings.php
@@ -49,6 +49,8 @@ class AuthCodesSettings extends BaseSettings implements SettingsInterface
             && !empty($this->settings['allow_delete'])
             || !empty($this->settings['auth_requires_auth_code'])
             && !empty($this->settings['allow_authentication'])
+            || !empty($this->settings['reset_password_requires_auth_code'])
+            && !empty($this->settings['allow_reset_password'])
         ) {
             if (empty($this->settings['auth_codes'])) {
                 throw new Exception(