+++ Changelog.md
@@ -1,6 +1,9 @@
 # Changelog - Releases
 
 ## (Unreleased)
+- Fix Auth Codes title on Authentication page
+- Add Reset password and Send Reset password endpoints 
+-
 
 ##  3.0.0 (11 July 2021)
 - Plugin code refactor